pyproject.toml
[project]
name = "ethyca_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "celery>=5.4.0",
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.6",
    "django-allauth[mfa]>=65.4.1",
    "django-anymail[sendgrid]>=12.0",
    "django-celery-beat>=2.7.0",
    "django-crispy-forms>=2.3",
    "django-environ>=0.12.0",
    "django-model-utils>=5.0.0",
    "django-storages[s3]>=1.14.5",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg[c,pool]>=3.2.5",
    "python-slugify>=8.0.4",
    "redis[hiredis]>=5.2.1",
    "sentry-sdk>=2.22.0",
    "whitenoise[brotli]>=6.9.0",
    "django-simple-history>=3.8.0",
    "django-localflavor>=4.0",
    "django-csp>=3.8",
    "django-permissions-policy>=4.25.0",
    "django-extensions>=3.2.3",
    "factory-boy>=3.3.2",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "django-debug-toolbar>=5.0.1",
    "django-stubs[compatible-mypy]>=5.1.3",
    "ipdb>=0.13.13",
    "pytest>=8.3.4",
    "pytest-django>=4.10.0",
    "pytest-sugar>=1.0.0",
    "watchfiles>=1.0.4",
    "werkzeug[watchdog]>=3.1.3",
    "pre-commit>=4.1.0",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["ethyca_project/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
line-length = 119
target-version = "py313"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "COM812", # The following rule may cause conflicts when used with the formatter: `COM812`.
            # To avoid unexpected behavior, we recommend disabling this rule
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true
